Server-ShortAnswers/requirements.txt
Flask
Flask-CORS
openai
python-dotenv
numpy
joblib
Keras
tensorflow-gpu
tensorflow

